+++ packages/plugins/GeoLocation/src/language.ts
@@ -6,6 +6,7 @@ const language: LanguageOption[] = [
     resources: {
       plugins: {
         geoLocation: {
+          markerText: 'Aktuelle Position',
           button: {
             tooltip: {
               placeLocationMarker: 'Eigene Position markieren',
@@ -25,6 +26,7 @@ const language: LanguageOption[] = [
     resources: {
       plugins: {
         geoLocation: {
+          markerText: 'Current location',
           button: {
             tooltip: {
               placeLocationMarker: 'Mark own location',